Removing domains
Removing domains that failed to retrieve lists. I'm not sure if I mentioned this but this list blocks 2,542,691 domains. You will have problems attempting to connect to the following social media sites: Twitter, Facebook(or anything owned by Facebook) and Instagram. The bypass is to whitelist the sites you'd like to visit. In addition to the following social media sites some mobile applications will not work regardless of the OS.